Final project - Wikipedia & translate API's
API modifications added:

   (1)    Google Translate or Bing Translate API - completed within the 'devices' class
   (2)    Wikipedia API  - completed within the 'devices' class
   (3)     Keywords displayed - completed within the 'devices' class 

-Notes: Instead of choosing to make different classes for each modification (i.e., the API's) I chose to work off the skeleton that my group and I had already created and use the 'devices' class, which worked around Lists to add and implement the modifications of the program (showing definitions and translating the keywords into German).

That being said, I also made some changes to the overall structure of the GUI, in which case I changed the colors and the initial messages that appear for the user. I made a few other changes as well, but most of my work was within the GUI and Devices classes.